luffyapi/requirements.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.4.0
cffi==1.13.2
cryptography==2.8
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.9
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
django-filter==2.2.0
django-formtools==2.2
django-import-export==2.0
django-redis==4.11.0
django-reversion==3.0.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
dnspython==1.16.0
et-xmlfile==1.0.1
eventlet==0.25.1
future==0.18.2
greenlet==0.4.15
httplib2==0.9.2
importlib-metadata==1.3.0
jdcal==1.4.1
kombu==4.6.7
MarkupPy==1.14
monotonic==1.5
more-itertools==8.0.2
odfpy==1.4.0
openpyxl==3.0.2
Pillow==7.0.0
pycparser==2.19
pycryptodomex==3.9.4
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
python-alipay-sdk==2.0.1
pytz==2019.3
PyYAML==5.2
qcloudsms-py==0.1.4
redis==3.3.11
six==1.13.0
tablib==0.14.0
vine==1.3.0
https://codeload.github.com/sshwsfc/xadmin/zip/django2
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
